form label to typescript
