+++ packages/design-system/src/components/FormLabel/FormLabel.test.tsx
@@ -73,7 +73,11 @@ describe('FormLabel', () => {
   });
 
   it('adds punctuation to requirementLabel when hint is also present', () => {
-    let props = { hint: 'Hint', requirementLabel: 'Optional' };
+    interface PropDef {
+      hint: React.ReactNode;
+      requirementLabel: React.ReactNode;
+    }
+    let props: PropDef = { hint: 'Hint', requirementLabel: 'Optional' };
     let wrapper = shallow(<FormLabel {...props}>{labelText}</FormLabel>);
 
     expect(wrapper).toMatchSnapshot();
@@ -95,8 +99,7 @@ describe('FormLabel', () => {
   });
 
   it('renders as a legend element', () => {
-    const props = { component: 'legend' };
-    const wrapper = shallow(<FormLabel {...props}>{labelText}</FormLabel>);
+    const wrapper = shallow(<FormLabel component="legend">{labelText}</FormLabel>);
 
     expect(wrapper.is('legend')).toBe(true);
   });
