+++ packages/design-system/src/components/FormLabel/FormLabel.tsx
@@ -1,10 +1,65 @@
 import InlineError from '../InlineError/InlineError';
-import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
 
-export class FormLabel extends React.PureComponent {
-  hint() {
+export interface FormLabelProps {
+  /**
+   * Label text or HTML.
+   */
+  children: React.ReactNode;
+  /**
+   * Additional classes to be added to the root element.
+   */
+  className?: string;
+  /** The root HTML element used to render the label */
+  component?: 'label' | 'legend';
+  /** Enable the error state by providing an error message. */
+  errorMessage?: React.ReactNode;
+  /**
+   * Additional classes to be added to the error message
+   */
+  errorMessageClassName?: string;
+  /**
+   * The ID of the error message applied to this field.
+   */
+  errorId?: string;
+  /**
+   * The ID of the field this label is for. This is used for the label's `for`
+   * attribute and any related ARIA attributes, such as for the error message.
+   */
+  fieldId?: string;
+  /**
+   * Additional hint text to display
+   */
+  hint?: React.ReactNode;
+  /**
+   * A unique `id` for the label element. Useful for referencing the label from
+   * other components with `aria-describedby`.
+   */
+  id?: string;
+  /**
+   * Set to `true` to apply the "inverse" theme
+   */
+  inversed?: boolean;
+  /**
+   * Text showing the requirement (ie. "Optional", or "Required").
+   * In most cases, this should be used to indicate which fields are optional.
+   * See the [form guidelines]({{root}}/guidelines/forms/) for more info.
+   */
+  requirementLabel?: React.ReactNode;
+  /**
+   * Additional classes to be added to the label text.
+   */
+  textClassName?: string;
+}
+
+export class FormLabel extends React.PureComponent<
+  React.ComponentPropsWithRef<'label'> & React.ComponentPropsWithRef<'legend'> & FormLabelProps,
+  any
+> {
+  static defaultProps = { component: 'label' };
+
+  hint(): React.ReactNode {
     const { hint } = this.props;
     let { requirementLabel } = this.props;
     if (!hint && !requirementLabel) return;
@@ -36,7 +91,7 @@ export class FormLabel extends React.PureComponent {
     );
   }
 
-  errorMessage() {
+  errorMessage(): React.ReactNode {
     if (this.props.errorMessage) {
       // Include fallback for errorId for usage outside of FormControl
       let errorId = null;
@@ -58,7 +113,7 @@ export class FormLabel extends React.PureComponent {
     }
   }
 
-  render() {
+  render(): React.ReactNode {
     const {
       fieldId,
       id,
@@ -88,57 +143,5 @@ export class FormLabel extends React.PureComponent {
     );
   }
 }
-
-FormLabel.defaultProps = { component: 'label' };
-FormLabel.propTypes = {
-  /**
-   * Label text or HTML.
-   */
-  children: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
-  /**
-   * Additional classes to be added to the root element.
-   */
-  className: PropTypes.string,
-  /** The root HTML element used to render the label */
-  component: PropTypes.oneOf(['label', 'legend']),
-  /** Enable the error state by providing an error message. */
-  errorMessage: PropTypes.node,
-  /**
-   * Additional classes to be added to the error message
-   */
-  errorMessageClassName: PropTypes.string,
-  /**
-   * The ID of the error message applied to this field.
-   */
-  errorId: PropTypes.string,
-  /**
-   * The ID of the field this label is for. This is used for the label's `for`
-   * attribute and any related ARIA attributes, such as for the error message.
-   */
-  fieldId: PropTypes.string,
-  /**
-   * Additional hint text to display
-   */
-  hint: PropTypes.node,
-  /**
-   * A unique `id` for the label element. Useful for referencing the label from
-   * other components with `aria-describedby`.
-   */
-  id: PropTypes.string,
-  /**
-   * Set to `true` to apply the "inverse" theme
-   */
-  inversed: PropTypes.bool,
-  /**
-   * Text showing the requirement (ie. "Optional", or "Required").
-   * In most cases, this should be used to indicate which fields are optional.
-   * See the [form guidelines]({{root}}/guidelines/forms/) for more info.
-   */
-  requirementLabel: PropTypes.node,
-  /**
-   * Additional classes to be added to the label text.
-   */
-  textClassName: PropTypes.string,
-};
 
 export default FormLabel;
