+++ packages/design-system/src/types/index.d.ts
@@ -3,7 +3,6 @@
  */
 export { default as Button } from './Button/Button';
 export { default as Dialog } from './Dialog/Dialog';
-export { default as FormLabel } from './FormLabel/FormLabel';
 export { default as HelpDrawer } from './HelpDrawer/HelpDrawer';
 export { default as HelpDrawerToggle } from './HelpDrawer/HelpDrawerToggle';
 export { default as MonthPicker, getMonthNames } from './MonthPicker/MonthPicker';
@@ -33,6 +32,7 @@ export * from './DateField';
 export * from './Dropdown';
 export * from './FilterChip';
 export * from './FormControl';
+export * from './FormLabel';
 export * from './InlineError';
 export * from './Spinner';
 export * from './TextField';
